include: gcc.h: Add STRUCT_PACKED_PRE and STRUCT_PACKED_POST
For different tool-chains, the use of "__packed" is different
for structure. For IAR, the __packed is used before the struct
keyword in a structure definition. For GCC, the __packed is
used at the end of the struct keyword in a structure definition.
To support different tool-chains, add STRUCT_PACKED_PRE and
STRUCT_PACKED_POST to support multiple too-chain to modify the
structure type.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>