include: gcc.h: Add ENUM_PACKED_PRE and ENUM_PACKED_POST
For different tool-chains, the use of "__packed" is different
for enumeration. For IAR, the __packed is used before the enum
keyword in a enumeration definition. For GCC, the __packed is
used at the end of the enum keyword in a enumeration
definition.
To support different tool-chains, add ENUM_PACKED_PRE and
ENUM_PACKED_POST to support multiple too-chain to modify
the enumeration type.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>